up: share photo

+++ function/G_revert_photo_path.py
@@ -3,6 +3,8 @@ import json
 import sqlite3
 import function.tools.config_io as config_io
 
+style = "personal"
+
 def move_picture(dl_path, source, target):
     if os.path.exists(f"{dl_path}/{target}") == True:
         if os.path.exists(f"{dl_path}/{source}") == False:
@@ -12,10 +14,17 @@ def move_picture(dl_path, source, target):
             print(f"[失败] {source} 无法恢复，来源 {target}，目录已有同名文件")
 
 def revert_photo_path():
-    dl_path = config_io.load('dl_path')
+    global style
+    if style == "personal":
+        dl_path = config_io.load('dl_path')
+    elif style == "share":
+        dl_path = config_io.load('share_dl_path')
     conn = sqlite3.connect('everphoto.db')
     c = conn.cursor()
-    c.execute('SELECT source, target FROM move_record')
+    if style == "personal": 
+        c.execute('SELECT source, target FROM move_record')
+    elif style == "share":
+        c.execute('SELECT source, target FROM shared_move_record')
     record = c.fetchall()
     conn.close()
     for source, target in record:
@@ -23,22 +32,31 @@ def revert_photo_path():
     
     conn = sqlite3.connect('everphoto.db')
     c = conn.cursor()
-    c.execute('DELETE FROM move_record')
+    if style == "personal":
+        c.execute('DELETE FROM move_record')
+    elif style == "share":
+        c.execute('DELETE FROM shared_move_record')
     conn.commit()
     conn.close()
 
-    for item in os.listdir(dl_path):
-        if os.path.isdir(f"{dl_path}/{item}") == True:
-            try:
-                os.rmdir(f"{dl_path}/{item}")
-            except:
-                print(f"[跳过] {item} 目录非空，跳过删除")
-                pass
+    if style == "personal":
+        for item in os.listdir(dl_path):
+            if os.path.isdir(f"{dl_path}/{item}") == True:
+                try:
+                    os.rmdir(f"{dl_path}/{item}")
+                except:
+                    print(f"[跳过] {item} 目录非空，跳过删除")
+                    pass
 
-def interface():
+def interface(type = "personal"):
+    global style
+    style = type
     os.system('cls')
     print("时光相册下载器")
-    print("当前进度：7. 恢复照片路径到整理前")
+    if style == "personal":
+        print("当前进度：7. 恢复照片路径到整理前")
+    elif style == "share":
+        print("当前进度：12. 恢复共享相册的路径到整理前")
     print("")
     print("注意事项：")
     print("1. 整理后已移动或重命名的文件无法恢复到之前的路径，将跳过")
@@ -46,14 +64,20 @@ def interface():
     print("正在检测是否整理过...")
     conn = sqlite3.connect('everphoto.db')
     c = conn.cursor()
-    c.execute("SELECT count(*) FROM sqlite_master WHERE type='table' AND name='move_record'")
+    if style == "personal":
+        c.execute("SELECT count(*) FROM sqlite_master WHERE type='table' AND name='move_record'")
+    elif style == "share":
+        c.execute("SELECT count(*) FROM sqlite_master WHERE type='table' AND name='shared_move_record'")
     if c.fetchone()[0] == 0:
         conn.close()
         print("未检测到整理记录，无需恢复")
         input("按回车键继续...")
         return
     else:
-        c.execute('SELECT source, target FROM move_record')
+        if style == "personal":
+            c.execute('SELECT source, target FROM move_record')
+        elif style == "share":
+            c.execute('SELECT source, target FROM shared_move_record')
         record = c.fetchall()
         conn.close()
         if len(record) == 0:
